pyproject.toml
[project]
name = "profilegen"
version = "1.0.0"
description = "GitHub profile README stats generator"
requires-python = ">=3.10"
dependencies = [
    "python-dateutil",
    "requests",
    "lxml",
    "python-dotenv",
]

[tool.ruff]
target-version = "py310"
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific
    "C4",    # flake8-comprehensions
    "PIE",   # flake8-pie
    "PT",    # flake8-pytest-style
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "FLY",   # flynt
    "PERF",  # perflint
    "FURB",  # refurb
    "LOG",   # flake8-logging
    "G",     # flake8-logging-format
    "T20",   # flake8-print (warn on print statements)
]
ignore = [
    "T201",   # allow print() - this is a CLI script
    "E501",   # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["profilegen"]

[tool.basedpyright]
pythonVersion = "3.10"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"
reportMissingImports = true
reportMissingTypeStubs = false

[dependency-groups]
dev = [
    "basedpyright>=1.39.3",
    "ruff>=0.15.12",
]
